Created a generic function getNewIndexFromPresets that returns the next available index.
